frontend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.32.2
Pillow>=10.3.0 